v1.03
Fixes:
- Removed dependencies on the GUI classes' Scene node.
- Allowed for the character to be anywhere in the image name, assuming
that it is the only character in the image name. Otherwise, it may be
placed with another arraylist where it would not belong.
- Organized code & fixed syntax. Added some comments explaining what
certain parts of the code does.
- Reduced repetition in object building, most changes will be made to a
one and only object that will be created in initialization.
- ...

Plans:
- Same as before.
- Fix some issues where Java will not free resources of objects no
longer in use. I have an idea on how to fix this and will implement when
possible.
- ...

This is also an attempt at fixing the .gitignore file as it was tracking
testing folders in which I do not want uploaded.