v1.11
- Everything has been re-instated minus the clear all option.
- I extended the save option to double as a layer order manager. I
included an image of what I envision it to be. If the user is dealing
with a lot of images and does not want to go back and re-do their
selection, they can re-order the images and continue editing or saving
the image.
- Added some icons. Free icons provided by RPGMaker.
- Re-organized code. Everything doesn't have a random name to it.
- Most static objects are only accessible through accessors or mutators
for security purposes. Returned to one CSS. Themes are not a big deal as
I am still working on functionality.