v1.16b
- Reinstated the alert window.
- Moved the threading used during the build process to its own class.
This along with conditionals were added to prevent the user from
spamming load requests causing the program to crash and will instead
allow them to cancel a build process, allowing them to load another
folder from scratch. Also, now we only have 1 thread doing all the work
which was the original intention.
- CacheSingle and CacheMultiple no longer require an "Object" as an
argument, but instead require their respective File[] and List<File[]>.
Object was used as an argument because I originally wanted the IO class
to return either one depending on the files it loaded. Instead I created
a seperate IO function that is specific to CacheMultiple while the
original was already tailored for CacheSingle.
- Because the Alert window doesn't stop the threading process but warns
you this commit is in beta, hence the v1.16b.  Also there is a check in
place to prevent the file dialog from appearing if the BuildThread is
alive.