v1.17
- Felt that 1.17 was more appropriate as I made changes overall to the
build process and gui elements.

Changes:
- I joined CacheSingle and CacheMultiple into one Cache class, the class
implements runnable as it is meant to be run on a separate thread. Also
makes it easier to interrupt it given the new alert window if the user
wants to load another folder.
- I changed accessibility to some GUI elements, nodes that weren't
accessed outside of the 'gui' package were set to protected, and others
who are accessed outside were left public. I split the DisplayHelp to
DisplayAbout, DisplayAbout now only contains the VBox with information
about the program. The help menu is now under the new DisplayMenuBar
class.
- Everything Menu/MenuBar related is now in the DisplayMenuBar class
- All GUI elements are accessed through accessors or mutators. No longer
are they accessed directly.
- Split IO to Read and Save. In Read, I combined the reading of the
directories for Single and Multiple, how it determines what type reading
it wants to do is done through a boolean.
- No changes were made to Save, except for being moved to its standalone
class.

Known Issues:
- I believe that in making the GUI nodes accessible through accessors
and mutators rather than directly resulted in breaking the CSS.

Plans:
- Continue on making the new save window.