v1.20b
- Renamed Classes for Organization Purposes
- Moved "SELECTED_IMAGES" from DrawPreview to CacheList. Makes more
sense this way and it leaves DrawPreview's draw() method to be modular.
- Draw() from DrawPreview now returns an Image and takes in an ArrayList
as an argument.
- Added a settings menu to enable disable Single/Multiple load types,
and further other settings as I add them.
- Removed the Menuitem to select directory load type, and moved to a pop
up when pressing the "Open" button and in the settings menu.
- Have the Save Menu half-working. You can disable images you no longer
want to use and the saved image will not generate that layer, however it
does not change the selected images in the main GUI. Nor can you change
the layer order yet, despite there being textfields with layer numbers.
- Moved initializations to separate class, but still called on startup.
For organization purposes - Will likely change back
- Read no longer returns an error if the user cancels the file chooser
dialog when using the Multiple load type.
- Read no longer needs an argument. It looks to the Settings to
determine load type before showing the file chooser dialogue.
- SaveFile() now takes in an image as an argument. This was how it was
orginally meant to be, but as it was not a big issue for the past few
updates I left it alone. Edited for independent use with the new Save
Menu.
- This update is considered to be in beta while the addition of the new
save menu is not fully implemented, it still works the same as before.