v1.20.1b
- Changed ArrayLists to LinkedLists. I've read that LinkedLists while
they use more memory, are easier on the Garbage Collector, currently in
testing, but still works as normal.
- Added a batch script, not working at the moment, I still need to
import the code to fit the program. I would have made it a separate
library, but I still need to edit the script as it is a quick and dirty
script to batch build a set of images.
- Remove license