v1.20.2b
- Made some changes to DisplayMenuSave, it no longer creates a new
stackpane, imageview, scene, stage, whenever the preview button is
clicked. Instead it has an inner class that is initialized on app start
up that is not created every call. Better optimized.
- Forgot to mention for 1.20.1b, I changed the way TryParse checks for
certain types of objects, rather than forcing an exception out which I
believe to be very inefficient and bad practice. I instead used
"instanceof" to determine whether or not the object is what is being
tested. It is much better and no longer does an object need to be
created and no longer does an exception need to be forced.