v1.20.7
- Integrated BatchScript into the main app. Through the File -> Open
menu the user can access the Batch script.
- While it integrated it is not fully functional in that it will stall
the application as it does not currently run in its own thread (so use
with caution).
- Made changes to the save and read classes to implement specific
methods for the batch code. Doesn't look too bad, but it still bloated
some of the code.
- Created a new Scene/Menu for Batch.
- Made the Open menu have a sub-menu, which now contains 2 menu options
for Generate Image and Batch. (Will fix this later, but for now it gets
the job done)
- Fixed how the progress bar is given progress in the Cache class. It
would be stuck at 100% and not incrementing by quarters.
- Added a new draw method specifically for Batch creation, therefore
renaming both draw methods.
- Removed Black border around Center Scrollpane and removed focus color,
the pane will just appear with imageviews once the user loads images.
- Added a separator between the center scrollpane and generated image
preview.
- Moved FileChooser and DirectoryChooser to a separate class. Both read
and save need access to both and instead of creating a new object for
both classes, they will access the same one. I think this means I will
need to eventually move Read and Save back into their own class, but as
I said before, for now its good and it works.
- Resized multiple scenes as objects in the scene appear cut off when
using this program in Linux.
- Added a check to the preview call, if a user has no images selected it
will no longer return an error, but do nothing instead. I will likely
re-instate a disabled Save/Preview button until images are either loaded
or are selected as I had in the earlier versions.
- Batch is now implemented and has been severely optimized and improved.
Despite that, I will need to go back and do further changes to get it to
work well as a Runnable object. For now it uses the JavaFX thread and
causes a stall.
- Added some extra benchmarking methods.
- Added ToImage, MI, MPC classes. They were originally methods in the
BatchScript class but I decided that they were better fit to be in their
separate classes. Although they are very specific to the batch use-case
they can definitely be altered later for other things.
- ToImage converts files to images. Taking a List<File[]> as an argument
- MI finds the largest array size from the List<T[]> which in this case
would be the List<File[]> from before. By finding the largest size it
will create a 2D array of size NxN with N being the largest size. Space
is definitly wasted, but an even 2D array would only cause issues.
Again, it works for now and thats all I want.
- MPC is finding the maximum possible combination of the images. This is
to preemptively calculate how many files the Batch's imageListToBuild
will need to take in. Both for speed and memory. Speed in that it won't
have to constantly rebuild its array and memory that each array
iteration won't stay in memory and cause a memory leak.
- Added a TryImage in the TryParse class
- Other changes and fixes